backend/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2025.11.12
cffi==1.15.1
charset-normalizer==3.4.4
cryptography==45.0.7
Deprecated==1.3.1
Django==3.2.25
django-cors-headers==4.1.0
django-environ==0.11.2
django-oauth-toolkit==2.3.0
djangorestframework==3.15.1
idna==3.10
jwcrypto==1.5.1
oauthlib==3.2.2
Pillow==9.5.0
psycopg2-binary==2.9.9
pycparser==2.21
python-dotenv==0.21.1
pytz==2025.2
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.7
wrapt==1.16.0
gunicorn==21.2.0
whitenoise==6.6.0
